refs #171 -- Drafted fix for Georgia and Indiana
Washington's birthday computation rules are confusing: what's on Wikipedia and what's published on the official calendars differ. I've put as many warnings as possible, and I'll rework on it when I'm done with the rest.

All the other Georgia holidays are ok.
Indiana other holidays to be tackled in another patch.