refs #171 -- Fixed Kansas
it looks like Kansas is not including the "non-shifting boxing day"...

Anyway, there are other issues for Kansas calendar...

According to wikipedia, Kansas only has all federal holidays, except the Columbus Day and Washington's Birthday.
Unfortunately, other sources mention XMas Eve for 2018, but not for other years. It looks like there are no clear rules. I'll stick this to the Wikipedia definitions, but... I'm a bit sad here...

Sources to lookup, if you want to help:

* http://www.admin.ks.gov/docs/default-source/ops/holidays/holidays2018.pdf
* http://www.kansas.gov/employee/documents/2017calendar.pdf
* https://publicholidays.us/kansas/2018-dates/
* https://en.wikipedia.org/wiki/Public_holidays_in_the_United_States#Kansas
* https://publicholidays.us/kansas/2018-dates/