refs #171 -- Fixed North Carolina
* Removed Columbus Day and Washington's Birthday,
* Revamped Christmas Day rules

WARNING: there's no clear documentation on how the XMas shifts work in North Carolina. At least one use-case (When XMas Day falls on SUN) is not covered by any official document available on the web, and the exact rules for the shift are hidden behind a 404 document. We may have to wait a bit and do some more research on this later on.