refs #171 -- Fixed South Carolina
* Removed the NoShiftBoxingDay, can't find a trace of it in any source,
* Removed Columbus Day,
* Removed Good Friday.

NOTE: If we're reading the 2017 Holiday official document, we can see
that XMas Eve is shifted on the previous FRI, instead of being moved to
the following MON. Is it the same bug we have in North Carolina? No
idea: there's no source, again, to define the computation rules for this
shift.