refs #171 -- Fixed Texas
Global revamp of the Texas module, including new parameters for state
holidays, that are meant to be "off-work" for most employees, but
anybody can swap the official days to their convenience.

It's documented in the module docstring, and taking advantage of the
inheritance of Python classes, anybody can adapt `Texas` calendar to
their needs.