Class docstring will give its name
I thought it was a bit sad that you had to give a `name` class property that was just for the sake of having a string to decribe the class. Describing the class means using a `docstring`.
Now the `name` property is deducted from the docstring, and all the Europe classes along with the "core" United States one are aligned with a more Pythonic way: a class should always have a docstring.