add JapanBank class The Bank of Japan lists all holidays and other dates closed on their public website, https://www.boj.or.jp/en/about/outline/holi.htm/.
In researching this dates, I found that the Japan calendar
also needed to be updated because of adjustments in 2019 and 2020 due
to the new emperor. Some fixed holidays were added and some variable
holidays got shifted around.

I have tested and verified all the Bank of Japan holidays match what
has been publicly posted for 2019 and 2020. I have also compared the
national holidays of Japan to what is available at
https://www.timeanddate.com/holidays/japan/.